slice refactor
