+++ lib/ui/button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tiki_style/tiki_style.dart';
 
-import 'upvoty_service.dart';
+import '../service.dart';
 
 class UpvotyWidgetButton extends StatelessWidget {
